backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pillow==10.4.0
numpy==1.26.4
pymongo==4.7.3
pydantic==2.7.4
starlette==0.37.2
opencv-python-headless==4.10.0.84
ultralytics==8.3.49
python-dotenv==1.0.1